// src/pages/challenges/challenges-types.ts
export enum EChallengeTask {
  wager = 'wager',
  play = 'play',
  win = 'win',
  winstreak = 'winstreak',
  primaryScore = 'primaryScore',
  secondaryScore = 'secondaryScore',
  completeChallenge = 'completeChallenge',
  dailyCheckIn = 'dailyCheckIn',
  deposit = 'deposit',
  depositNTimes = 'depositNTimes',
  invite = 'invite',
  inviteXAndDeposit = 'inviteXAndDeposit',
  inviteAndDepositX = 'inviteAndDepositX',
  xpPoint = 'xpPoint',
  xpLevel = 'xpLevel',
  winning = 'winning',
  inviteAndPlay = 'inviteAndPlay',
  inviteAndLevelUp = 'inviteAndLevelUp',
};

export enum EChallengeCashFlag {
  all = 'all',
  cash = 'cash',
  coin = 'coin',
  ticket = 'ticket',
};

export enum EChallengeCashFlagForCrypto {
  all = 'all',
  JRX = 'JRX',
  jrx = 'coin',
  ticket = 'ticket',
};

export enum EChallengeTournamentType {
  all = 'all',
  bracketTournament = 'bracketTournament',
  asyncPlay = 'asyncPlay',
  syncPlay = 'syncPlay',
  asyncFFASinglePlayer = 'asyncFFASinglePlayer',
  multiEntry = 'multiEntry',
  asyncFFAWithWinQualifier = 'asyncFFAWithWinQualifier',
  singlePlay = 'singlePlay',
};